pass aligned types arguments by reference to make 32bit msvc happy fixed some compiler warnings
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@2384 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ features/include/pcl/features/impl/fpfh.hpp
@@ -192,7 +192,7 @@ pcl::FPFHEstimation<PointInT, PointNT, PointOutT>::computeFeature (PointCloudOut
     {
         // Special case: When a feature must be computed at every point, there is no need for a neighborhood search
         for (size_t idx = 0; idx < indices_->size (); ++idx)
-            spfh_indices.insert (idx);
+            spfh_indices.insert ((int) idx);
     }
 
     // Initialize the arrays that will store the SPFH signatures
